Launch a Shiny app for scouting

Launch a Shiny app for scouting

ov_scouter(
  dvw,
  video_file,
  court_ref,
  season_dir,
  auto_save_dir,
  scoreboard = TRUE,
  ball_path = FALSE,
  review_pane = TRUE,
  playlist_display_option = "dv_codes",
  scouting_options = ov_scouter_options(),
  default_scouting_table = ov_default_scouting_table(),
  compound_table = ov_default_compound_table(),
  shortcuts = ov_default_shortcuts(),
  launch_browser = TRUE,
  prompt_for_files = interactive(),
  ...
)

Arguments

dvw

string or datavolley: either the path to a dvw or ovs file or a datavolley object (e.g. as returned by dv_create(). Passing the file name (not the datavolley object) is required if any extra arguments are passed via .... dvw can also be an object as saved by ov_scouter() in ovs format. If dvw is "demo", the app will be started with a demonstration data set

video_file

string: optionally, the path to the video file. If not supplied (or NULL) the video file specified in the dvw file will be used. video_file can also be a URL (including a YouTube URL or video ID). NOTE that in this case, court_ref must be provided, it can't yet be entered interactively in the app

court_ref

data.frame or string: data.frame with the court reference (as returned by ovideo::ov_shiny_court_ref()) or the path to the rds file containing the output from this

season_dir

string: optional path to a directory with other dvw/ovs files from this season

auto_save_dir

string: optional path to a directory where the dvw will be saved automatically after each rally

scoreboard

logical: if TRUE, show a scoreboard in the top-right of the video pane

ball_path

logical: if TRUE, show the ball path on the court inset diagram. Note that this will slow the app down slightly

review_pane

logical: if TRUE, entry popups will be accompanied by a small video pane that shows a loop of the video of the action in question

playlist_display_option

string: what to show in the plays table? Either "dv_codes" (scouted codes) or "commentary" (a plain-language interpretation of the touches)

scouting_options

list: a named list with entries as per ov_scouter_options()

default_scouting_table

tibble: the table of scouting defaults (skill type and evaluation)

compound_table

tibble: the table of compound codes

shortcuts

list: named list of keyboard shortcuts, as returned by ov_default_shortcuts()

launch_browser

logical: if TRUE, launch the app in the system's default web browser (passed to shiny::runApp()'s launch.browser parameter)

prompt_for_files

logical: if dvw was not specified, prompt the user to select the dvw file

...

: extra parameters passed to datavolley::dv_read() (if dvw is a provided as a string) and/or to the shiny server and UI functions

Examples

if (FALSE) {
  ov_scouter("demo")
}