Scouting options

Scouting options

Usage

ov_scouting_options(
  end_convention = "actual",
  nblockers = TRUE,
  default_nblockers = NA,
  transition_sets = FALSE,
  attacks_by = "codes",
  team_system = "SHM3",
  setter_dump_code = "PP",
  second_ball_attack_code = "P2",
  overpass_attack_code = "PR",
  default_scouting_table = ov_default_scouting_table(),
  compound_table = ov_default_compound_table()
)

Arguments

end_convention

string: either "actual" or "intended". Is the end coordinate of an attack or serve the actual end location (where the ball contacted the floor or out of bounds area), or the intended one. The actual might differ from the intended if there is a block touch or the ball hit the net. If "actual", and a block touch is recorded, then the end location of the attack will not be used for the dig location (the dig location will be missing)

nblockers

logical: scout the number of blockers on each attack?

default_nblockers

integer: if nblockers is TRUE, what number of blockers should we default to? If NA, no default

transition_sets

logical: scout sets in transition? If FALSE, just the endpoint of each attack (i.e. the dig) and the subsequent counter-attack are scouted

attacks_by

string: "codes" (X5, V5, etc) or "tempo" (high, medium, quick)

team_system

string: the assumed system that teams are using to assign e.g. passing and hitting responsibilities

"SHM3" - a setter-hitter-middle rotation, with 3 passers (the libero and two outside hitters)

setter_dump_code

string: the attack combination code for a setter dump

second_ball_attack_code

string: the attack combination code for a second-ball attack

overpass_attack_code

string: the attack combination code for an attack on an overpass

default_scouting_table

tibble: the table of scouting defaults (skill type and evaluation)

compound_table

tibble: the table of compound codes

Value

A named list